requirements.txt
certifi==2021.10.8
cffi==1.14.3
charset-normalizer==2.0.9
greenlet==0.4.13
idna==3.3
python-dotenv==0.19.2
readline==6.2.4.1
requests==2.26.0
urllib3==1.26.7
